Fix Laravel CI workflow

+++ app/Modules/POS/Order/api.php
@@ -8,5 +8,5 @@ Route::group(['middleware' => ['auth:api']], function () {
     Route::post('orders', [OrderController::class, 'store']);
     Route::get('orders/{identifier}', [OrderController::class, 'show']);
     Route::patch('orders/{id}/status', [OrderController::class, 'updateStatus']);
-    Route::delete('orders/{id}', [OrderController::class, 'destroy']);
+    Route::delete('orders/{id}', [OrderController::class, 'destroy'])->middleware('permission:pos.void');
 });
